requirements.txt
# Core dependencies
flask>=2.3.0
openai>=1.0.0
numpy>=1.21.0
tqdm>=4.64.0

# NLP and ML
torch>=1.13.0
transformers>=4.25.0
sentence-transformers>=2.2.0
nltk>=3.7.0

# Web scraping and data processing
requests>=2.28.0
beautifulsoup4>=4.11.0

# System utilities
psutil>=5.9.0

# Optional for better performance

argparse~=1.4.0
pillow~=11.1.0
huggingface-hub~=0.33.0